Replace styledOmitProps with omitProps
Why? Wrapping emotion styled is not compatible with emotions babel plugin
