+++ packages/components/src/DatePicker/components/CalendarMonth/CalendarMonth.js
@@ -1,13 +1,14 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { themeGet } from 'styled-system';
+import styled from '@emotion/styled';
 import { css } from '@emotion/core';
 import { Box, Text } from '../../../';
 import CalendarWeekdays from '../CalendarWeekdays';
 import CalendarDays from '../CalendarDays';
-import styledOmitProps from '../../../styledOmitProps';
+import omitProps from '../../../omitProps';
 
-const MonthWrapper = styledOmitProps(Box, { omit: ['monthsToDisplay', 'stacked'] })`
+const MonthWrapper = styled(Box, omitProps(['monthsToDisplay', 'stacked']))`
   text-align: center;
   padding: 0 ${themeGet('space.4')};
   width: ${props => `${100 / props.monthsToDisplay}%`};
