+++ packages/components/src/omitProps/index.js
@@ -1,2 +1,2 @@
-export { default } from './styledOmitProps';
+export { default } from './omitProps';
 export { VALID_STYLED_SYSTEM_PROPS } from './constants';
